extract common data structures
